backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
python-dateutil==2.8.2